Moves test classes from /google to /utilities/gdrive
