+++ collect_app/src/test/java/org/odk/collect/android/utilities/gdrive/GoogleAccountsManagerTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.google;
+package org.odk.collect.android.utilities.gdrive;
 
 import android.content.Intent;
 
@@ -12,7 +12,6 @@ import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
 import org.odk.collect.android.preferences.PreferenceKeys;
-import org.odk.collect.android.utilities.gdrive.GoogleAccountsManager;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
